requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
anyio==3.6.2
async-generator==1.10
async-timeout==4.0.2
attrs==22.2.0
autopep8==2.0.1
certifi==2022.12.7
charset-normalizer==2.1.1
click==8.1.3
exceptiongroup==1.1.0
fastapi==0.89.1
fetcher-for-zjuer==1.0.0
frozenlist==1.3.3
h11==0.14.0
httptools==0.5.0
idna==3.4
loguru==0.6.0
msgpack==1.0.4
multidict==6.0.4
nonebot-adapter-onebot==2.2.1
nonebot2==2.0.0rc3
outcome==1.2.0
packaging==23.0
Pillow==9.4.0
pip-autoremove==0.10.0
pycodestyle==2.10.0
pydantic==1.10.4
pydub==0.25.1
PyExecJS==1.5.1
pygtrie==2.5.0
pypinyin==0.48.0
PySocks==1.7.1
python-dotenv==0.21.1
pytz==2022.7.1
requests==2.28.2
selenium==4.8.0
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
starlette==0.22.0
tomli==2.0.1
tomlkit==0.11.6
tqdm==4.64.1
trio==0.22.0
trio-websocket==0.9.2
typing_extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
uvloop==0.17.0
watchfiles==0.18.1
webdriver-manager==3.8.5
websockets==10.4
wsproto==1.2.0
yarl==1.8.2
